+++ src/rust/src/x509/ocsp_resp.rs
@@ -834,6 +834,7 @@ pub(crate) fn create_ocsp_response(
         private_key.clone(),
         hash_algorithm.clone(),
         py.None().into_bound(py),
+        None,
         &tbs_bytes,
     )?;