+++ tests/x509/test_x509.py
@@ -48,7 +48,10 @@ from ..hazmat.primitives.fixtures_ec import EC_KEY_SECP256R1
 from ..hazmat.primitives.fixtures_rsa import (
     RSA_KEY_2048_ALT,
 )
-from ..hazmat.primitives.test_ec import _skip_curve_unsupported
+from ..hazmat.primitives.test_ec import (
+    _skip_curve_unsupported,
+    _skip_deterministic_ecdsa_unsupported,
+)
 from ..hazmat.primitives.test_rsa import rsa_key_512, rsa_key_2048
 from ..utils import (
     load_nist_vectors,
@@ -3617,6 +3620,71 @@ class TestCertificateBuilder:
             x509.DNSName("cryptography.io"),
         ]
 
+    def test_build_cert_with_deterministic_ecdsa_signature(self, backend):
+        _skip_curve_unsupported(backend, ec.SECP256R1())
+        _skip_deterministic_ecdsa_unsupported(backend)
+
+        private_key = ec.generate_private_key(ec.SECP256R1())
+
+        not_valid_before = datetime.datetime(2002, 1, 1, 12, 1)
+        not_valid_after = datetime.datetime(2030, 12, 31, 8, 30)
+
+        builder = (
+            x509.CertificateBuilder()
+            .serial_number(777)
+            .issuer_name(x509.Name([]))
+            .subject_name(x509.Name([]))
+            .public_key(private_key.public_key())
+            .not_valid_before(not_valid_before)
+            .not_valid_after(not_valid_after)
+        )
+        cert1 = builder.sign(
+            private_key,
+            hashes.SHA256(),
+            backend,
+            ecdsa_deterministic=True,
+        )
+        cert2 = builder.sign(
+            private_key,
+            hashes.SHA256(),
+            backend,
+            ecdsa_deterministic=True,
+        )
+        cert_nondet = builder.sign(private_key, hashes.SHA256(), backend)
+        cert_nondet2 = builder.sign(
+            private_key, hashes.SHA256(), backend, ecdsa_deterministic=False
+        )
+
+        assert cert1.signature == cert2.signature
+        assert cert1.signature != cert_nondet.signature
+        assert cert_nondet.signature != cert_nondet2.signature
+
+        private_key.public_key().verify(
+            cert1.signature,
+            cert1.tbs_certificate_bytes,
+            ec.ECDSA(hashes.SHA256()),
+        )
+
+    def test_sign_deterministic_wrong_key_type(self, rsa_key_2048, backend):
+        not_valid_before = datetime.datetime(2002, 1, 1, 12, 1)
+        not_valid_after = datetime.datetime(2030, 12, 31, 8, 30)
+        builder = (
+            x509.CertificateBuilder()
+            .serial_number(777)
+            .issuer_name(x509.Name([]))
+            .subject_name(x509.Name([]))
+            .public_key(rsa_key_2048.public_key())
+            .not_valid_before(not_valid_before)
+            .not_valid_after(not_valid_after)
+        )
+        with pytest.raises(TypeError):
+            builder.sign(
+                rsa_key_2048,
+                hashes.SHA256(),
+                backend,
+                ecdsa_deterministic=True,
+            )
+
     def test_build_cert_with_bmpstring_universalstring_name(
         self, rsa_key_2048: rsa.RSAPrivateKey, backend
     ):
@@ -4786,6 +4854,61 @@ class TestCertificateSigningRequestBuilder:
         assert isinstance(basic_constraints.value, x509.BasicConstraints)
         assert basic_constraints.value.ca is True
         assert basic_constraints.value.path_length == 2
+
+    def test_build_ca_request_with_deterministic_ec(self, backend):
+        _skip_curve_unsupported(backend, ec.SECP256R1())
+        _skip_deterministic_ecdsa_unsupported(backend)
+
+        private_key = ec.generate_private_key(ec.SECP256R1())
+
+        builder = x509.CertificateSigningRequestBuilder().subject_name(
+            x509.Name([])
+        )
+        csr1 = builder.sign(
+            private_key,
+            hashes.SHA256(),
+            backend,
+            ecdsa_deterministic=True,
+        )
+        csr2 = builder.sign(
+            private_key,
+            hashes.SHA256(),
+            backend,
+            ecdsa_deterministic=True,
+        )
+
+        csr_nondet = builder.sign(
+            private_key,
+            hashes.SHA256(),
+            backend,
+        )
+
+        csr_nondet2 = builder.sign(
+            private_key,
+            hashes.SHA256(),
+            backend,
+        )
+
+        assert csr1.signature == csr2.signature
+        assert csr1.signature != csr_nondet.signature
+        assert csr_nondet.signature != csr_nondet2.signature
+        private_key.public_key().verify(
+            csr1.signature,
+            csr1.tbs_certrequest_bytes,
+            ec.ECDSA(hashes.SHA256()),
+        )
+
+    def test_csr_deterministic_wrong_key_type(self, rsa_key_2048, backend):
+        builder = x509.CertificateSigningRequestBuilder().subject_name(
+            x509.Name([])
+        )
+        with pytest.raises(TypeError):
+            builder.sign(
+                rsa_key_2048,
+                hashes.SHA256(),
+                backend,
+                ecdsa_deterministic=True,
+            )
 
     @pytest.mark.supported(
         only_if=lambda backend: backend.ed25519_supported(),