+++ CHANGELOG.rst
@@ -69,6 +69,10 @@ Changelog
   involving small buffers.
 * Added :func:`~cryptography.hazmat.primitives.serialization.ssh_key_fingerprint`
   for computing fingerprints of SSH public keys.
+* Added support for deterministic ECDSA signing via the new keyword-only argument
+  ``ecdsa_deterministic`` in :meth:`~cryptography.x509.CertificateBuilder.sign`,
+  :meth:`~cryptography.x509.CertificateRevocationListBuilder.sign`
+  and :meth:`~cryptography.x509.CertificateSigningRequestBuilder.sign`.
 
 .. _v44-0-3: